// DataSharing/src/model/nca.model.ts
import { NcaDataTypesModel } from './nca.datatypes.model';
import { NcaLambdaZModel } from './nca.lambdaz.model';
import { NcaMappingsModel } from './mappings/nca.mappings.model'
import { NcaOptionsModel } from './nca.options.model';
import { NcaResultsModel } from './nca.results.model';
import { NcaSettingsModel } from './nca.settings.model';

export class NcaModel
{
     ncaDataTypesModel: NcaDataTypesModel;
     ncaLambdaZModel: NcaLambdaZModel;
     ncaMappingsModel: NcaMappingsModel;
     ncaOptionsModel: NcaOptionsModel;
     ncaResultsModel: NcaResultsModel;
     ncaSettingsModel: NcaSettingsModel;
    
}